fix(api-gateway): cap in-flight keys, dry-run metric
In-flight zones read max_keys only in the periodic sweep. Between
sweeps, a per-IP zone could grow without limit from distinct keys. This
is the same hole that W2 closed for rate zones.

InFlightZone::gate now refuses a new key when the zone is at max_keys.
The request gets a 429 with kind=max_keys and Retry-After equal to the
prune interval. In dry-run mode, the request is served without a gate.
The sweep opens admission again. Both zone kinds keep the key count in
an AtomicU64, so the hot path does not walk every DashMap shard.

Dry-run mode only wrote a warn log for each request. Operators had no
metric to graph. Add the throttling.dry_run_rejections counter with the
same zone/kind attributes as throttling.rejections. It is a separate
instrument, so the existing series does not change. The three
log_dry_run_* helpers become one record_dry_run function.

Move the throttling unit tests to throttling_tests.rs. Add tests for
the cap, for concurrent admission, and for the new metric.

BREAKING CHANGE: in_flight_limit_zones.max_keys now rejects new keys
past the cap. Before, the zone tracked them until the next sweep.
